+++ zscaler/zwa/v2_client.go
@@ -141,7 +141,7 @@ func getHTTPClient(l logger.Logger, rateLimiter *rl.RateLimiter, cfg *Configurat
 	}
 
 	if cfg.ZWA.Testing.DisableHttpsCheck {
-		transport.TLSClientConfig = &tls.Config{InsecureSkipVerify: false}
+		transport.TLSClientConfig = &tls.Config{InsecureSkipVerify: true}
 		l.Printf("[INFO] HTTPS certificate validation is disabled (testing mode).")
 	}